requirements.txt
djangorestframework
python-social-auth
django-extensions